Crate zng_txt

String type optimized for sharing.

Crate

This crate is part of the zng project.

Macros

formatx

Creates a Txt by formatting using the format_args! syntax.

Structs

Txt

Text string type, can be one of multiple internal representations, mostly optimized for sharing and one for editing.

Enums

TxtRepr

Identifies how a Txt is currently storing the string data.

Traits

ToTxt

A trait for converting a value to a Txt.